Attempted Fix on Railcraft IC2Classic Support
onStructureChanged changes: IsMaster is if the master is being called, this one is either always true or always false, but has no indication if the structure is valid or not.
Meaning we only care about the master if we do changes, but the "isComplete" (is formed) is the indicator if the EU Stuff should be applied or not.

changes to add/drop: Basically trust the internal state, if that one is not trustworthy then you have other problems...
Also just drop duplicated calls.

If a structure can be reformed while still being valid then we have different issues again.

Note: this patch attempt was just done with some skimming of the data and what made sense at the time.

A few things I couldn't figure out:
The master is expected to be "index = 0" in the "getSubTiles" list. As per IC2Exp specifications.